Scale to metric and adjust component sizes
I do not quite understand what XSize and YSize should be in Macrofab
format.  I've tried several approaches, and the one that works best is
to just put the longest footprint dimension in XSize.  This gets most
parts right, but still results in an unacceptable error rate of about 5%
of parts.

Fortunately, what counts for manufacturing is X,Y and rotation, and
those we have them right.